Updated SQLite model, UI polish and page number in gallery 's title when asked for new collection

+++ Virtual tourist/PinMO+CoreDataClass.swift
@@ -26,7 +26,7 @@ public class PinMO: NSManagedObject {
     }
     
     func setAlbumValues(photoPages: Int, totalPhotos: Int) {
-        self.photoPages = Int16(photoPages)
+        self.photoPages = Int64(photoPages)
         self.totalPhotos = Int64(totalPhotos)
     }
 }
